requirements.txt
accelerate==1.3.0
controlnet-aux==0.0.9
diffusers==0.32.2
ExifRead==3.0.0
ffmpeg==1.4
ffmpeg-python==0.2.0
GPUtil==1.4.0
moviepy==1.0.3
onnxruntime-gpu==1.20.1
opencv-python==4.11.0.86
Panda3D==1.10.15
pilgram==1.2.1
pillow==10.2.0
PyMatting==1.1.13
PyQt6==6.8.1
pyqtgraph==0.13.7
qimage2ndarray==1.10.0
rawpy==0.24.0
realesrgan==0.3.0
rembg==2.0.62
timm==0.6.7
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
transformers==4.48.3
tqdm==4.66.5